CodeSpec had too much going on, split defs of Mod, Func, FuncStmt, etc to new file and other functions to Chunk.Code where they belong
