+++ code/drasil-code/Language/Drasil/Chunk/Code.hs
@@ -1,13 +1,16 @@
 {-# LANGUAGE TemplateHaskell #-}
 module Language.Drasil.Chunk.Code (
     CodeIdea(..), CodeChunk(..), codeType, codevar, codefunc, quantvar, 
-    quantfunc, ccObjVar, ConstraintMap, constraintMap, physLookup, sfwrLookup,
-    programName, funcPrefix
+    quantfunc, ccObjVar, codevars, codevars', funcResolve, varResolve, 
+    ConstraintMap, constraintMap, physLookup, sfwrLookup, programName, 
+    funcPrefix
   ) where
 
 import Control.Lens ((^.),makeLenses,view)
 
 import Language.Drasil
+import Database.Drasil (ChunkDB, symbResolve)
+import Language.Drasil.Development (dep, names')
 
 import Language.Drasil.Chunk.CodeQuantity (HasCodeType(ctyp), CodeQuantityDict, 
   cqw, implCQD)
@@ -15,6 +18,7 @@ import Language.Drasil.Printers (symbolDoc, toPlainName)
 
 import GOOL.Drasil (CodeType(Object))
 
+import Data.List (nub)
 import qualified Data.Map as Map
 import Text.PrettyPrint.HughesPJ (render)
 
@@ -71,6 +75,18 @@ ccObjVar c1 c2 = checkObj (codeType c1)
           (Concat [symbol c1 Implementation, Label ".", symbol c2 
           Implementation]) (getUnit c2) 
         checkObj _ = error "First CodeChunk passed to ccObjVar must have Object type"
+
+-- | Get a list of CodeChunks from an equation
+codevars :: Expr -> ChunkDB -> [CodeChunk]
+codevars e m = map (varResolve m) $ dep e
+
+-- | Get a list of CodeChunks from an equation (no functions)
+codevars' :: Expr -> ChunkDB -> [CodeChunk]
+codevars' e m = map (varResolve m) $ nub $ names' e
+
+funcResolve, varResolve :: ChunkDB -> UID -> CodeChunk
+funcResolve m x = quantfunc $ symbResolve m x
+varResolve  m x = quantvar  $ symbResolve m x
 
 type ConstraintMap = Map.Map UID [Constraint]
