+++ code/drasil-code/Language/Drasil/CodeSpec.hs
@@ -5,24 +5,23 @@ import Language.Drasil
 import Database.Drasil (ChunkDB, SystemInformation(SI), symbResolve,
   _authors, _constants, _constraints, _datadefs, _definitions, _inputs,
   _outputs, _quants, _sys, _sysinfodb, sampleData)
-import Language.Drasil.Development (dep, names', namesRI)
+import Language.Drasil.Development (dep, namesRI)
 import Theory.Drasil (DataDefinition, qdFromDD)
 
 import Language.Drasil.Chunk.Code (CodeChunk, CodeIdea(codeChunk), 
-  ConstraintMap, programName, codevar, quantvar, quantfunc, funcPrefix, 
-  codeName, constraintMap)
+  ConstraintMap, programName, codevar, quantvar, funcPrefix, codeName, 
+  codevars, codevars', funcResolve, varResolve, constraintMap)
 import Language.Drasil.Chunk.CodeDefinition (CodeDefinition, qtov, qtoc, 
   codeEquat)
 import Language.Drasil.Chunk.CodeQuantity (HasCodeType(ctyp))
-import Language.Drasil.Code.Code (spaceToCodeType)
 import Language.Drasil.Code.CodeQuantityDicts (inFileName, inParams, consts)
-import Language.Drasil.Code.DataDesc (DataDesc, getInputs)
-import Language.Drasil.Printers (toPlainName)
+import Language.Drasil.Mod (Func(..), FuncData(..), FuncDef(..), Mod(..), Name, fname, getFuncParams, packmod, 
+  prefixFunctions)
 
 import GOOL.Drasil (CodeType)
 
 import Control.Lens ((^.))
-import Data.List (nub, delete, (\\))
+import Data.List (nub, (\\))
 import qualified Data.Map as Map
 import Data.Maybe (mapMaybe)
 
@@ -211,8 +210,6 @@ defaultChoices = Choices {
   auxFiles = [SampleInput]
 }
 
-type Name = String
-
 -- medium hacks ---
 relToQD :: ExprRelat c => ChunkDB -> c -> QDefinition
 relToQD sm r = convertRel sm (r ^. relat)
@@ -220,59 +217,6 @@ relToQD sm r = convertRel sm (r ^. relat)
 convertRel :: ChunkDB -> Expr -> QDefinition
 convertRel sm (BinaryOp Eq (C x) r) = ec (symbResolve sm x) r
 convertRel _ _ = error "Conversion failed"
-
-data Mod = Mod Name String [Func]
-
-packmod :: Name -> String -> [Func] -> Mod
-packmod n = Mod (toPlainName n)
-
-data DMod = DMod [Name] Mod
-     
-data Func = FCD CodeDefinition
-          | FDef FuncDef
-          | FData FuncData
-
-funcQD :: QDefinition -> Func
-funcQD qd = FCD $ qtoc qd 
-
-funcData :: Name -> String -> DataDesc -> Func
-funcData n desc d = FData $ FuncData (toPlainName n) desc d
-
-funcDef :: (Quantity c, MayHaveUnit c) => Name -> String -> [c] -> Space -> 
-  Maybe String -> [FuncStmt] -> Func  
-funcDef s desc i t returnDesc fs = FDef $ FuncDef (toPlainName s) desc 
-  (map quantvar i) (spaceToCodeType t) returnDesc fs 
-
-data FuncData where
-  FuncData :: Name -> String -> DataDesc -> FuncData
-  
-data FuncDef where
-  FuncDef :: Name -> String -> [CodeChunk] -> CodeType -> Maybe String -> 
-    [FuncStmt] -> FuncDef
- 
-data FuncStmt where
-  FAsg :: CodeChunk -> Expr -> FuncStmt
-  FAsgIndex :: CodeChunk -> Integer -> Expr -> FuncStmt
-  FFor :: CodeChunk -> Expr -> [FuncStmt] -> FuncStmt
-  FWhile :: Expr -> [FuncStmt] -> FuncStmt
-  FCond :: Expr -> [FuncStmt] -> [FuncStmt] -> FuncStmt
-  FRet :: Expr -> FuncStmt
-  FThrow :: String -> FuncStmt
-  FTry :: [FuncStmt] -> [FuncStmt] -> FuncStmt
-  FContinue :: FuncStmt
-  FDec :: CodeChunk -> FuncStmt
-  FProcCall :: Func -> [Expr] -> FuncStmt
-  -- slight hack, for now
-  FAppend :: Expr -> Expr -> FuncStmt
-  
-($:=) :: (Quantity c, MayHaveUnit c) => c -> Expr -> FuncStmt
-v $:= e = FAsg (quantvar v) e
-
-ffor :: (Quantity c, MayHaveUnit c) => c -> Expr -> [FuncStmt] -> FuncStmt
-ffor v = FFor (quantvar  v)
-
-fdec :: (Quantity c, MayHaveUnit c) => c -> FuncStmt
-fdec v  = FDec (quantvar v)
 
 asVC :: Func -> QuantityDict
 asVC (FDef (FuncDef n _ _ _ _ _)) = implVar n (nounPhraseSP n) (Variable n) Real
@@ -302,11 +246,6 @@ getAdditionalVars chs ms = map codevar (inFileName
         constsVar (Store Bundled) = [consts]
         constsVar _ = []
         funcParams (Mod _ _ fs) = concatMap getFuncParams fs
-
-getFuncParams :: Func -> [CodeChunk]
-getFuncParams (FDef (FuncDef _ _ ps _ _ _)) = ps
-getFuncParams (FData (FuncData _ _ d)) = getInputs d
-getFuncParams (FCD _) = []
 
 -- name of variable/function maps to module name
 type ModExportMap = Map.Map String String
@@ -345,50 +284,6 @@ clsDefMap cs@CSI {
     ++ getDerivedCls chs ds
     ++ getConstraintsCls chs (getConstraints (cMap cs) ins)
     ++ getInputFormatCls chs extIns
-
-fstdecl :: ChunkDB -> [FuncStmt] -> [CodeChunk]
-fstdecl ctx fsts = nub (concatMap (fstvars ctx) fsts) \\ nub (concatMap (declared ctx) fsts) 
-  where
-    fstvars :: ChunkDB -> FuncStmt -> [CodeChunk]
-    fstvars _  (FDec cch) = [cch]
-    fstvars sm (FAsg cch e) = cch:codevars' e sm
-    fstvars sm (FAsgIndex cch _ e) = cch:codevars' e sm
-    fstvars sm (FFor cch e fs) = delete cch $ nub (codevars' e sm ++ concatMap (fstvars sm) fs)
-    fstvars sm (FWhile e fs) = codevars' e sm ++ concatMap (fstvars sm) fs
-    fstvars sm (FCond e tfs efs) = codevars' e sm ++ concatMap (fstvars sm) tfs ++ concatMap (fstvars sm) efs
-    fstvars sm (FRet e) = codevars' e sm
-    fstvars sm (FTry tfs cfs) = concatMap (fstvars sm) tfs ++ concatMap (fstvars sm ) cfs
-    fstvars _  (FThrow _) = [] -- is this right?
-    fstvars _  FContinue = []
-    fstvars sm (FProcCall _ l) = concatMap (`codevars` sm) l
-    fstvars sm (FAppend a b) = nub (codevars a sm ++ codevars b sm)
-
-    declared :: ChunkDB -> FuncStmt -> [CodeChunk]
-    declared _  (FDec cch) = [cch]
-    declared _  (FAsg _ _) = []
-    declared _  FAsgIndex {} = []
-    declared sm (FFor _ _ fs) = concatMap (declared sm) fs
-    declared sm (FWhile _ fs) = concatMap (declared sm) fs
-    declared sm (FCond _ tfs efs) = concatMap (declared sm) tfs ++ concatMap (declared sm) efs
-    declared _  (FRet _) = []
-    declared sm (FTry tfs cfs) = concatMap (declared sm) tfs ++ concatMap (declared sm) cfs
-    declared _  (FThrow _) = [] -- is this right?
-    declared _  FContinue = []
-    declared _  (FProcCall _ _) = []
-    declared _  (FAppend _ _) = []
-       
-fname :: Func -> Name       
-fname (FCD cd) = codeName cd
-fname (FDef (FuncDef n _ _ _ _ _)) = n
-fname (FData (FuncData n _ _)) = n 
-
-prefixFunctions :: [Mod] -> [Mod]
-prefixFunctions = map (\(Mod nm desc fs) -> Mod nm desc $ map pfunc fs)
-  where pfunc f@(FCD _) = f
-        pfunc (FData (FuncData n desc d)) = FData (FuncData (funcPrefix ++ n) 
-          desc d)
-        pfunc (FDef (FuncDef n desc a t rd f)) = FDef (FuncDef (funcPrefix ++ n)
-          desc a t rd f)
 
 getDerivedInputs :: [DataDefinition] -> [QDefinition] -> [Input] -> [Const] ->
   ChunkDB -> [QDefinition]
@@ -519,14 +414,6 @@ xs `subsetOf` ys = all (`elem` ys) xs
 getConstraints :: ConstraintMap -> [CodeChunk] -> [Constraint]
 getConstraints cm cs = concat $ mapMaybe (\c -> Map.lookup (c ^. uid) cm) cs
 
--- | Get a list of CodeChunks from an equation
-codevars :: Expr -> ChunkDB -> [CodeChunk]
-codevars e m = map (varResolve m) $ dep e
-
--- | Get a list of CodeChunks from an equation (no functions)
-codevars' :: Expr -> ChunkDB -> [CodeChunk]
-codevars' e m = map (varResolve m) $ nub $ names' e
-
 -- | Get a list of CodeChunks from an equation, where the CodeChunks are correctly parameterized by either Var or Func
 codevarsandfuncs :: Expr -> ChunkDB -> ModExportMap -> [CodeChunk]
 codevarsandfuncs e m mem = map resolve $ dep e
@@ -541,7 +428,3 @@ constraintvarsandfuncs (Range _ ri) m mem = map resolve $ nub $ namesRI ri
           | Map.member (funcPrefix ++ x) mem = funcResolve m x
           | otherwise = varResolve m x
 constraintvarsandfuncs _ _ _ = []
-
-funcResolve, varResolve :: ChunkDB -> UID -> CodeChunk
-funcResolve m x = quantfunc $ symbResolve m x
-varResolve  m x = quantvar  $ symbResolve m x
